Remove our copy of maas_signal.py
distutils edits shebang lines of things when it installs them, so since we're
installing with python3, we can't have python2 scripts. There is really no
reason for us to keep a copy of maas_signal.py and its dependencies in our
tree, so we just use the one installed by maas.