Update juju state and single installer UI processing
juju state now has a separate machine class that makes it
easier to inspect and perform basic operations on the required
UI parameters.

The single UI has been updated to make use of the latest changes in
juju state. In a perfect world the multi-install UI that was previously
in place should not be impacted, however, im sure there are some small
changes we need to make within that code path.

Signed-off-by: Adam Stokes <adam.stokes@ubuntu.com>