Import MaaS images in a more sane way
So we'd like to run m-i-p-f ourselves, because if we run it through the API
there is no way to get feedback. However, MaaS uses a celery task to reap the
images after they're imported, which the API calls for us but running it
directly via the CLI doesn't. If we don't reap the images, MaaS won't let us
create the bootstrap node. So we call the celery job manually.