Around we go again back to LXC
According to the openstack-charmers team trying to put
all openstack bits (except compute) on 1 machine is unsupported.
Primarily due to the fact that HAProxy plays a huge part in the
exposing and serving of services such as Horizon.

Knowing this now forces us back to using LXC containers in a single
installation for everything.

Signed-off-by: Adam Stokes <adam.stokes@ubuntu.com>