Cosmetic cleanups and logic re-work
Some of the cosmetics with how the nodeview is displayed
have changed from just a plain text string to utilizing
columns. This gives the UI a more fluid feel even during
resizing of terminal emulators.

The logic was re-worked a bit to do away with worrying about
unallocated vs allocated machines in a single install. Since
those instances are created on-demand this code isn't necessary
any longer. It's kept in for review but currently commented
out.

Signed-off-by: Adam Stokes <adam.stokes@ubuntu.com>