support injecting local copy of OpenStack installer
This is mainly for developers. Previously there wasn't
and easy way to test latest code changes for the single
installer since it was always pulling from the defined ppa
via cloud-init.

This adds ability to copy a local copy of the installer package
to be installed into the container for testing.

Fixes #213

Signed-off-by: Adam Stokes <adam.stokes@ubuntu.com>