remove custom network alterations to host containers network.
By default the host container is going to use lxcbr0 with the 10.0.4.x network
since the host _system_ is using lxcbr0 on the 10.0.3.x network. With this
setup we can then followup with a static route for the host system interface
to point directly to 10.0.4.x without having to deal with any additional
bridge modifications.

Signed-off-by: Adam Stokes <adam.stokes@ubuntu.com>