Remove juju/stable ppa
To many regressions between stable releases. We are going to
concentrate on what's in the official archive (1.20.11 at the
time of commit). 1.20.14 is due out soon and that has been tested
with our installer and known to work.

If wanting to test new juju versions that can still be done like:

`sudo openstack-install --extra-ppa ppa:juju/stable`

Signed-off-by: Adam Stokes <adam.stokes@ubuntu.com>