Fix proxy handling with openstack-git enabled
Removes hardcoded proxy value in the template for neutron charms, which
should fix installs from openstack-git that are not behind a proxy.

Sends proxy values to charms in charmconf, which should not be necessary
permanently. This currently will still not work for any firewall that
blocks access to git.openstack.org using the git port, however.
That will depend on a fix for this bug:
https://bugs.launchpad.net/charm-helpers/+bug/1472416

Signed-off-by: Michael McCracken <mike.mccracken@canonical.com>