Update logic handling of configuration during run of install and status
The status screen should always parse an existing config file or fail.
When transitioning from the install screen over to the status screen
the config object was getting overwritten since there was no pre-existing
config object available. We set a default location for the --config-file
parameter to make sure our configuration options are loaded into the
status screen correctly. You still have to ability to pass additional
configuration options via the config file or cli.

The same applies for the 'uninstall' option in `openstack-install` as
it requires the `install_type` config option when deciding on what
to cleanup.

Signed-off-by: Adam Stokes <adam.stokes@ubuntu.com>