Check if a deploy is completed and skip deployment actions
Before we were just re-running the charm processors blindly no
matter if a deploy was complete or not. Even though it didn't
cause an issue there was a visual misrepresentation of the state
of the status screen.

Now we check for a completed deploy and update the UI to be 'Ready'