Remove ability to install new MAAS
I think we should remove this and if users would like to try out MAAS
provide some sort of pre-built LXC image that contains an already
configured MAAS server.

It hasn't been tested/addressed in quite awhile and the code is just
sitting around getting stale.

I've already begun the process of getting a container built with MAAS
so that users can just download the image and load it up on whatever
machine they want.

Signed-off-by: Adam Stokes <adam.stokes@ubuntu.com>