Pass --insecure pollinate via --curl-opts
This is to fix an issue where users may have
self-signed certificates as sort of a mitm proxy.

I prefer to not blindly accept all certs but in the
interest of making things work with minimal configuration
we'll use this for now.

We may revisit this and attach it to some sort of
--insecure flag of our own.

Closes #669

Signed-off-by: Adam Stokes <adam.stokes@ubuntu.com>