Change have_nextbranch to available_sources
We needed a way to define charms that are available in
some, all, any of particular branches in order to be made
available depending on the requirements.

For example, currently, the 'tempest' charm is only
available in a /next branch and not in the charmstore.

This feature allows us to be more specific as to what
remote sources are available for said charm.

Signed-off-by: Adam Stokes <adam.stokes@ubuntu.com>