Use an observable for handling events
Currently, if an exception occurs we use a popup box
to let the user know that an error happened. However,
in some cases this may be hidden behind the fact that
the update alarms are still running. Meaning, it could
potentionally hide the overlay with another UI event
like updating the progress status view.

The problem this observer solves is it allows us to
call various functions based on the Event() emitted.

This code shows how that is possible when an exception
happens as it will remove the eventloops alarms and
allow the exception error message to stay seen in the
UI.

The other minor thing is the reason why I did this was
because I don't want to stop the eventloop entirely b/c
we'll still need access to keyboard commands, like (q)
in order for the user to quit out of the ui if a problem
occurs.

Signed-off-by: Adam Stokes <adam.stokes@ubuntu.com>