Use string for missing juju machine id
Fixes crash in display when machine can not be created for a unit, or is
not ready yet.

When a machine can't be found, the default machine id of -1 is returned,
which is eventually passed as-is to a Text() constructor, which fails.

Nothing other than display and comparisons uses this value, so it's ok
to use something arbitrary as long as it's not something juju would
actually assign to a machine.

Signed-off-by: Mike McCracken <mike.mccracken@canonical.com>