Drop tornado for asyncio
Use whats provided with python 3.4 and drops an additional dependency
of tornado. The original reason for tornado was to make use of the
more advanced asyncio loop but now that urwid 1.3.1 is backported
to trusty in our ppa we can skip that and use whats provided in
python 3.4. urwid 1.3.1 is available by default in xenial and above.

Signed-off-by: Adam Stokes <adam.stokes@ubuntu.com>