async: clean up threads on quit
Shutdown was hanging after the urwid loop stopped, due to threads in the
thread pool looping forever, or just for a long time.

Async launched threads need to occasionally check if they need to kill
themselves.

Added async.shutdown to cleanly shutdown the executor and signal an
event that the threads check occasionally.

Added async.sleep_until to provide a sleep() workalike that checks every
.1 seconds for the shutdown Event and raises an exception if it is set.
Occasionally this is used with a timeout of 0 to just check for the
event once.

These changes should quit cleanly within a minimum of time - although
sometimes it will still stall while waiting for e.g. a remote execution
of a post-processing script, as the checks are only inline.