Add 'source' option to glance-simplestreams-sync
The cloud-installer/simplestreams-testing PPA has a copy of
simplestreams for trusty that includes the progress reporting code we
need. It had been installed unconditionally by the GSS charm but as that
charm is now in use by others, that has caused problems.

Instead we have added a 'source' config option in this MP:
https://code.launchpad.net/~mikemc/charms/trusty/glance-simplestreams-sync/add-source-config/+merge/282806
this allows us to just set it here and others will not use it by default.

Signed-off-by: Michael McCracken <mike.mccracken@canonical.com>